adding a CollectionsParserMixin
This mixin will hold useful functions for handling collections from
refresh_parsers.
