+++ app/models/manageiq/providers/kubernetes/container_manager/refresh_parser.rb
@@ -133,19 +133,7 @@ module ManageIQ::Providers::Kubernetes
       end
     end
 
-    def process_collection(collection, key, &block)
-      @data[key] ||= []
-      collection.each { |item| process_collection_item(item, key, &block) }
-    end
-
-    def process_collection_item(item, key)
-      @data[key] ||= []
-
-      new_result = yield(item)
-
-      @data[key] << new_result
-      new_result
-    end
+    include CollectionsParserMixin
 
     def map_labels(model_name, labels)
       ContainerLabelTagMapping.map_labels(@label_tag_mapping, model_name, labels)
